Refac: move n2v_utils
