+++ src/careamics/lightning/modules/n2v_utils/pixel_manipulation.py
