+++ src/careamics/lightning/modules/n2v_utils/struct_mask_parameters.py
