+++ tests/compat/transforms/test_old_manipulate_n2v.py
@@ -5,7 +5,7 @@ import torch
 from careamics.compat.transforms.n2v_manipulate import N2VManipulate as N2VManipulateNpy
 from careamics.config.augmentations import N2VManipulateConfig
 from careamics.config.support import SupportedPixelManipulation
-from careamics.lightning.lightning_modules.n2v_utils import N2VManipulate
+from careamics.lightning.modules.n2v_utils import N2VManipulate
 
 
 @pytest.mark.parametrize(
